Crate reth_blockchain_tree

Implementation of a tree-like structure for blockchains.

The BlockchainTree can validate, execute, and revert blocks in multiple competing sidechains. This structure is used for Reth’s sync mode at the tip instead of the pipeline, and is the primary executor and validator of payloads sent from the consensus layer.

Blocks and their resulting state transitions are kept in-memory until they are persisted.

Feature Flags

test-utils: Export utilities for testing

Re-exports

pub use blockchain_tree::BlockchainTree;

pub use block_indices::BlockIndices;

pub use chain::AppendableChain;

pub use config::BlockchainTreeConfig;

pub use externals::TreeExternals;

pub use shareable::ShareableBlockchainTree;

pub use block_buffer::BlockBuffer;

Modules

block_buffer

Buffer of not executed blocks.

block_indices

Implementation of BlockIndices related to super::BlockchainTree

blockchain_tree

Implementation of BlockchainTree

chain

A chain in a BlockchainTree.

config

Blockchain tree configuration

externals

Blockchain tree externals.

metrics

Common blockchain tree metrics.

noop

Implementation of Tree traits that does nothing.

shareable

Wrapper around BlockchainTree that allows for it to be shared.

Structs

BundleStateData

Structure that owns the relevant data needs to be a BundleStateDataProvider

BundleStateDataRef

Structure that combines references of required data to be a BundleStateDataProvider.